pyproject.toml
[project]
name = "torch-training-script"
version = "0.1.0"
description = "PyTorch object detection training pipeline with multiclass support"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]

[project.optional-dependencies]
# Core dependencies needed for all operations
core = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.21.0",
    "pillow>=9.0.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
]

# Training-specific dependencies (includes core)
train = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.21.0",
    "pillow>=9.0.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "tqdm>=4.64.0",
    "torchinfo>=1.8.0",
    "tensorboard>=2.10.0",
]

# Evaluation-specific dependencies (includes core)
eval = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.21.0",
    "pillow>=9.0.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "pycocotools>=2.0.0",
    "matplotlib>=3.5.0",
]

# Hyperparameter optimization (includes train)
sweep = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.21.0",
    "pillow>=9.0.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "tqdm>=4.64.0",
    "torchinfo>=1.8.0",
    "tensorboard>=2.10.0",
    "optuna>=2.10.0,<3.0.0",  # Version constraint from hydra-optuna-sweeper
    "hydra-optuna-sweeper>=1.2.0",
]

# ONNX Vision Service module (lightweight, no PyTorch)
vision-service = [
    "viam-sdk>=0.24.0",
    "onnxruntime>=1.16.0",
    "numpy>=1.21.0",
    "pillow>=9.0.0",
    "pyinstaller>=6.0.0",
    "google-api-python-client",
]

# Development dependencies
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]

# All dependencies (for convenience)
all = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.21.0",
    "pillow>=9.0.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "tqdm>=4.64.0",
    "torchinfo>=1.8.0",
    "pycocotools>=2.0.0",
    "matplotlib>=3.5.0",
    "tensorboard>=2.11.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

